Alternando o comando double para float

+++ exercicio3.java
@@ -3,13 +3,13 @@ public class exercicio3 {
 
 	public static void main(String[] args) {
 		
-		double b, B, h, area;
+		float b, B, h, area;
 		
-		b = 6.0;
-		B = 8.0;
-		h = 5.0;
+		b = 6f;
+		B = 8f;
+		h = 5f;
 		
-		area = (b + B) / 2.0 * h;
+		area = (float) ((b + B) / 2.0 * h);
 		
 		System.out.println(area);
